for #2408, move ExtractorRuleDefinitionEntityLoaderTest from sharding-core-parse-test to sharding-core-parse-common
