+++ sharding-core/sharding-core-parse/sharding-core-parse-common/src/test/java/org/apache/shardingsphere/core/parse/rule/AllRuleTests.java
@@ -18,11 +18,15 @@
 package org.apache.shardingsphere.core.parse.rule;
 
 import org.apache.shardingsphere.core.parse.rule.jaxb.loader.RuleDefinitionFileConstantTest;
+import org.apache.shardingsphere.core.parse.rule.jaxb.loader.extractor.ExtractorRuleDefinitionEntityLoaderTest;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 import org.junit.runners.Suite.SuiteClasses;
 
 @RunWith(Suite.class)
-@SuiteClasses(RuleDefinitionFileConstantTest.class)
+@SuiteClasses({
+        RuleDefinitionFileConstantTest.class,
+        ExtractorRuleDefinitionEntityLoaderTest.class
+})
 public final class AllRuleTests {
 }
